+++ components/Setting.tsx
@@ -1,0 +1,5 @@
+import {OtherPage} from './OtherPage';
+
+export const Setting = () => {
+  return <OtherPage title="設定"></OtherPage>;
+};
